+++ packages/library/src/rpc-request-coalescer.ts
@@ -1,4 +1,4 @@
-import type { IRpcTransport } from '@solana/rpc-transport';
+import type { IRpcTransport } from '@solana/rpc-types';
 
 type CoalescedRequest = {
     readonly abortController: AbortController;
