+++ packages/rpc-transport/src/__tests__/json-rpc-test.ts
@@ -1,10 +1,9 @@
-import { IRpcApi, Rpc, RpcRequest } from '@solana/rpc-types';
+import { IRpcApi, IRpcTransport, Rpc, RpcRequest } from '@solana/rpc-types';
 
 import { createJsonRpc } from '../json-rpc';
 import { SolanaJsonRpcError } from '../json-rpc-errors';
 import { createJsonRpcMessage } from '../json-rpc-message';
 import { getNextMessageId } from '../json-rpc-message-id';
-import { IRpcTransport } from '../transports/transport-types';
 
 jest.mock('../json-rpc-message-id');
