+++ packages/rpc-transport/src/transports/http/__tests__/http-transport-abort-test.ts
@@ -1,6 +1,6 @@
+import { IRpcTransport } from '@solana/rpc-types';
 import fetchMock from 'jest-fetch-mock-fork';
 
-import { IRpcTransport } from '../../transport-types';
 import { createHttpTransport } from '../http-transport';
 
 describe('createHttpTransport and `AbortSignal`', () => {
