+++ packages/rpc-transport/src/transports/transport-types.ts
@@ -1,14 +1,5 @@
 import { RpcWebSocketConnection } from './websocket/websocket-connection';
 
-type RpcTransportConfig = Readonly<{
-    payload: unknown;
-    signal?: AbortSignal;
-}>;
-
-export interface IRpcTransport {
-    <TResponse>(config: RpcTransportConfig): Promise<TResponse>;
-}
-
 type RpcWebSocketTransportConfig = Readonly<{
     payload: unknown;
     signal: AbortSignal;
